refactor(button): Replace old button module with refactored button
