--- modules/_labs/button/react/index.ts
@@ -1,9 +1,0 @@
-export {default as Button, default as beta_Button, ButtonProps} from './lib/Button';
-
-export {default as DeleteButton, DeleteButtonProps} from './lib/DeleteButton';
-export {default as deprecated_Button, DeprecatedButtonProps} from './lib/deprecated_Button';
-export {default as DropdownButton, DropdownButtonProps} from './lib/DropdownButton';
-export {default as HighlightButton, HighlightButtonProps} from './lib/HighlightButton';
-export {default as IconButton, IconButtonProps} from './lib/IconButton';
-export {default as TextButton, TextButtonProps} from './lib/TextButton';
-export * from './lib/types';
