--- modules/_labs/button/react/lib/DropdownButton.tsx
@@ -1,53 +1,0 @@
-import * as React from 'react';
-import {caretDownIcon} from '@workday/canvas-system-icons-web';
-import {GrowthBehavior} from '@workday/canvas-kit-react-common';
-import {DropdownButtonVariant, ButtonIconPosition} from './types';
-import {ButtonContainer, ButtonLabel, ButtonLabelIcon} from './parts';
-import {getButtonColors} from './Button';
-
-export interface DropdownButtonProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
-    GrowthBehavior {
-  /**
-   * The variant of the Button.
-   * @default DropdownButtonVariant.Secondary
-   */
-  variant?: DropdownButtonVariant;
-  /**
-   * The size of the Button.
-   * @default 'medium'
-   */
-  size?: 'medium' | 'large';
-  /**
-   * The ref to the button that the styled component renders.
-   */
-  buttonRef?: React.Ref<HTMLButtonElement>;
-}
-
-const DropdownButton = ({
-  variant = DropdownButtonVariant.Secondary,
-  size = 'medium',
-  buttonRef,
-  children,
-  ...elemProps
-}: DropdownButtonProps) => {
-  return (
-    <ButtonContainer colors={getButtonColors(variant)} size={size} ref={buttonRef} {...elemProps}>
-      <ButtonLabel>{children}</ButtonLabel>
-      <ButtonLabelIcon
-        size={size}
-        icon={caretDownIcon}
-        iconPosition={ButtonIconPosition.Right}
-        dropdown={true}
-      />
-    </ButtonContainer>
-  );
-};
-
-DropdownButton.Variant = DropdownButtonVariant;
-DropdownButton.Size = {
-  Medium: 'medium',
-  Large: 'large',
-} as const;
-
-export default DropdownButton;
