--- modules/_labs/button/react/package.json
@@ -1,52 +1,0 @@
-
-{
-  "name": "@workday/canvas-kit-labs-react-button",
-  "version": "0.0.0",
-  "description": "A collection of various Canvas Buttons",
-  "author": "Workday, Inc. (https://www.workday.com)",
-  "license": "Apache-2.0",
-  "main": "dist/commonjs/index.js",
-  "module": "dist/es6/index.js",
-  "sideEffects": false,
-  "types": "dist/es6/index.d.ts",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Workday/canvas-kit/tree/master/modules/_labs/button/react"
-  },
-  "files": [
-    "dist/",
-    "lib/",
-    "index.ts"
-  ],
-  "scripts": {
-    "watch": "yarn build:es6 -w",
-    "clean": "rimraf dist && rimraf .build-info && mkdirp dist",
-    "build:cjs": "tsc -p tsconfig.cjs.json",
-    "build:es6": "tsc -p tsconfig.es6.json",
-    "build:rebuild": "npm-run-all clean build",
-    "build": "npm-run-all --parallel build:cjs build:es6",
-    "depcheck": "node ../../../../utils/check-dependencies-exist.js"
-  },
-  "keywords": [
-    "canvas",
-    "canvas-kit",
-    "react",
-    "components",
-    "workday",
-    "button"
-  ],
-  "peerDependencies": {
-    "react": ">= 16.8 < 17"
-  },
-  "dependencies": {
-    "@emotion/core": "^10.0.21",
-    "@emotion/is-prop-valid": "^0.8.2",
-    "@emotion/styled": "^10.0.17",
-    "@workday/canvas-kit-labs-react-core": "^3.4.0",
-    "@workday/canvas-kit-react-common": "^3.4.0",
-    "@workday/canvas-kit-react-core": "^3.4.0",
-    "@workday/canvas-kit-react-icon": "^3.4.0",
-    "@workday/canvas-system-icons-web": "^1.0.20",
-    "@workday/design-assets-types": "^0.2.4"
-  }
-}
