--- modules/_labs/button/react/tsconfig.es6.json
@@ -1,9 +1,0 @@
-
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "declaration": true,
-    "outDir": "dist/es6",
-    "tsBuildInfoFile": "./.build-info/tsconfig.es6.tsbuildinfo"
-  }
-}
