--- modules/_labs/button/react/tsconfig.json
@@ -1,5 +1,0 @@
-
-{
-  "extends": "../../../../tsconfig.json",
-  "exclude": ["node_modules", "ts-tmp", "dist", "spec", "stories"]
-}
