+++ modules/button/react/lib/IconButton.tsx
@@ -1,33 +1,29 @@
 import * as React from 'react';
-import {styled} from '@workday/canvas-kit-labs-react-core';
-import isPropValid from '@emotion/is-prop-valid';
-import {IconButtonVariant, IconButtonSize} from './types';
-import {iconButtonStyles} from './ButtonStyles';
-import {getButtonStyle} from './utils';
-import {colors} from '@workday/canvas-kit-react-core';
+import {colors, spacing, borderRadius} from '@workday/canvas-kit-react-core';
+import {focusRing} from '@workday/canvas-kit-react-common';
 import {SystemIcon} from '@workday/canvas-kit-react-icon';
-import {focusRing, mouseFocusBehavior} from '@workday/canvas-kit-react-common';
 import {CanvasSystemIcon} from '@workday/design-assets-types';
-import {ClassNames, CSSObject} from '@emotion/core';
+import {IconButtonVariant, ButtonColors} from './types';
+import {ButtonContainer} from './parts';
 
 export interface IconButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-  /**
-   * The type of the IconButton.
-   * @default IconButtonVariant.Circle
-   */
-  variant: IconButtonVariant;
   /**
    * The accessibility label to indicate the action triggered by clicking the IconButton.
    */
   'aria-label': string;
   /**
-   * The size of the IconButton.
-   * @default IconButtonSize.Medium
+   * The type of the IconButton.
+   * @default IconButtonVariant.Circle
    */
-  size?: IconButtonSize;
+  variant?: IconButtonVariant;
   /**
-   * If true, toggle the IconButton on.
-   * @default false
+   * The size of the IconButton.
+   * @default 'medium'
+   */
+  size?: 'small' | 'medium';
+  /**
+   * The toggled state of the button. If defined as a boolean, then it manages the toggled state: on (`true`) or off (`false`).
+   * @default undefined
    */
   toggled?: boolean;
   /**
@@ -44,227 +40,187 @@ export interface IconButtonProps extends React.ButtonHTMLAttributes<HTMLButtonEl
   onToggleChange?: (toggled: boolean | undefined) => void;
 }
 
-function getFillSelector(fillColor: string): CSSObject {
-  return {
-    '&:focus span .wd-icon-fill, &:hover span .wd-icon-fill, span .wd-icon-fill': {
-      fill: fillColor,
-    },
-  };
-}
+const IconButton = ({
+  variant = IconButtonVariant.Circle,
+  size = 'medium',
+  buttonRef,
+  onToggleChange,
+  icon,
+  toggled,
+  children,
+  ...elemProps
+}: IconButtonProps) => {
+  const isInitialMount = React.useRef(true);
 
-function getBackgroundSelector(fillColor: string): CSSObject {
-  return {
-    '&:hover span .wd-icon-background, span .wd-icon-background': {
-      fill: fillColor,
-    },
-  };
-}
-
-function getAccentSelector(fillColor: string): CSSObject {
-  return {
-    '&:focus span .wd-icon-accent, &:hover span .wd-icon-accent, span .wd-icon-accent': {
-      fill: fillColor,
-    },
-  };
-}
-
-export const iconButtonIdentifier = 'wdc-ckr-icon-button';
-
-export const IconButtonCon = styled('button', {
-  shouldForwardProp: prop => isPropValid(prop) && prop !== 'size',
-})<IconButtonProps>(
-  iconButtonStyles.styles,
-  ({variant}) => getButtonStyle(iconButtonStyles, variant),
-  ({size, variant}) => {
-    switch (size) {
-      default:
-      case IconButtonSize.Medium:
-        return {
-          margin: variant === IconButtonVariant.Plain ? '-8px' : undefined,
-          ...iconButtonStyles.variants!.sizes.medium,
-        };
-      case IconButtonSize.Small:
-        return {
-          margin: variant === IconButtonVariant.Plain ? '-6px' : undefined,
-          ...iconButtonStyles.variants!.sizes.small,
-        };
-    }
-  },
-  ({variant, toggled}) => {
-    if (!toggled) {
-      return {};
-    }
-    switch (variant) {
-      case IconButtonVariant.CircleFilled:
-      case IconButtonVariant.SquareFilled:
-      case IconButtonVariant.Circle:
-      case IconButtonVariant.Square:
-      default: {
-        return {
-          backgroundColor: colors.blueberry400,
-          ...getFillSelector(colors.frenchVanilla100),
-          ...getAccentSelector(colors.blueberry400),
-          ...getBackgroundSelector(colors.frenchVanilla100),
-          '&:focus&:hover, &:focus, &:active, &:active:hover': {
-            ...getFillSelector(colors.frenchVanilla100),
-            ...getAccentSelector(colors.blueberry400),
-            ...getBackgroundSelector(colors.frenchVanilla100),
-            backgroundColor: colors.blueberry500,
-          },
-          '&:not([disabled])': {
-            '&:focus, &:focus:active': {
-              backgroundColor: colors.blueberry500,
-              ...(toggled ? focusRing(2, 2) : {}),
-            },
-          },
-          '&:hover, &:active': {
-            ...getFillSelector(colors.frenchVanilla100),
-            ...getAccentSelector(colors.blueberry400),
-            ...getBackgroundSelector(colors.frenchVanilla100),
-            backgroundColor: colors.blueberry500,
-          },
-          '&:disabled, &:active:disabled, &:focus:disabled, &:hover:disabled': {
-            backgroundColor: colors.blueberry100,
-            ...getFillSelector(colors.blueberry300),
-            ...getAccentSelector(colors.frenchVanilla100),
-            ...getBackgroundSelector(colors.blueberry300),
-          },
-          ...mouseFocusBehavior({
-            '&:focus:active': {
-              ...getFillSelector(colors.frenchVanilla100),
-              ...getAccentSelector(colors.blueberry400),
-              ...getBackgroundSelector(colors.frenchVanilla100),
-              backgroundColor: `${colors.blueberry500} !important`,
-            },
-          }),
-        };
+  // Only call onToggleChange on update - not on first mount
+  React.useEffect(() => {
+    if (isInitialMount.current) {
+      isInitialMount.current = false;
+    } else {
+      if (toggled && typeof onToggleChange === 'function') {
+        onToggleChange(toggled);
       }
-
-      case IconButtonVariant.Plain:
-        return {
-          backgroundColor: 'transparent',
-          ...getFillSelector(colors.blueberry400),
-          ...getAccentSelector(colors.frenchVanilla100),
-          ...getBackgroundSelector(colors.blueberry400),
-          '&:focus:hover, &:focus, &:active, &:active:hover': {
-            ...getFillSelector(colors.blueberry400),
-            ...getAccentSelector(colors.frenchVanilla100),
-            ...getBackgroundSelector(colors.blueberry400),
-          },
-          '&:not([disabled]):focus': {
-            ...getFillSelector(colors.blueberry400),
-            ...getAccentSelector(colors.frenchVanilla100),
-            ...getBackgroundSelector(colors.blueberry400),
-            ...(toggled ? focusRing(2, 0) : {}),
-          },
-          '&:disabled, &:active:disabled, &:focus:disabled, &:hover:disabled': {
-            ...getFillSelector(colors.blueberry200),
-            ...getAccentSelector(colors.frenchVanilla100),
-            ...getBackgroundSelector(colors.blueberry200),
-          },
-        };
-
-      case IconButtonVariant.Inverse:
-      case IconButtonVariant.InverseFilled:
-        return {
-          ...getFillSelector(colors.blueberry400),
-          ...getAccentSelector(colors.frenchVanilla100),
-          ...getBackgroundSelector(colors.blueberry400),
-          backgroundColor: colors.frenchVanilla100,
-          '&:focus': {
-            backgroundColor: colors.frenchVanilla100,
-            ...getFillSelector(colors.blueberry400),
-            ...getAccentSelector(colors.frenchVanilla100),
-            ...getBackgroundSelector(colors.blueberry400),
-          },
-          '&:focus&:hover, &:active, &:active:hover': {
-            backgroundColor: colors.frenchVanilla100,
-            ...getFillSelector(colors.blueberry400),
-            ...getAccentSelector(colors.frenchVanilla100),
-            ...getBackgroundSelector(colors.blueberry400),
-          },
-          '&:focus:active': {
-            backgroundColor: colors.frenchVanilla100,
-          },
-
-          '&:hover': {
-            backgroundColor: colors.frenchVanilla100,
-          },
-          '&:not([disabled])': {
-            '&:focus': {
-              backgroundColor: colors.frenchVanilla100,
-              ...(toggled
-                ? focusRing(2, 2, true, false, 'currentColor', colors.frenchVanilla100)
-                : {}),
-            },
-            '&:focus:active': {
-              backgroundColor: colors.frenchVanilla100,
-            },
-          },
-          '&:disabled, &:active:disabled, &:focus:disabled, &:hover:disabled': {
-            backgroundColor: 'rgba(255,255,255,0.75)',
-            ...getFillSelector(colors.blueberry400),
-            ...getAccentSelector(colors.frenchVanilla100),
-            ...getBackgroundSelector(colors.blueberry400),
-          },
-          ...mouseFocusBehavior({
-            '&:focus:active': {
-              backgroundColor: `${colors.frenchVanilla100} !important`,
-            },
-          }),
-        };
     }
+  }, [toggled, onToggleChange]);
+
+  const containerStyles = {
+    padding: 0,
+    margin: variant === IconButtonVariant.Plain ? '-8px' : undefined,
+    minWidth: size === 'small' ? spacing.l : spacing.xl, // min-width is set so buttons don't collapse in IE11
+    width: size === 'small' ? spacing.l : spacing.xl,
+    height: size === 'small' ? spacing.l : spacing.xl,
+    borderRadius:
+      variant === IconButtonVariant.Square || variant === IconButtonVariant.SquareFilled
+        ? borderRadius.m
+        : borderRadius.circle,
+    ['& .wd-icon']: {
+      display: 'inline-block',
+      verticalAlign: 'middle',
+      width: size === 'small' ? '20px' : undefined,
+      height: size === 'small' ? '20px' : undefined,
+    },
+  };
+
+  return (
+    <ButtonContainer
+      colors={getIconButtonColors(variant, toggled)}
+      size={size}
+      ref={buttonRef}
+      fillIcon={toggled}
+      extraStyles={containerStyles}
+      aria-pressed={toggled}
+      {...elemProps}
+    >
+      {icon ? <SystemIcon icon={icon} /> : children}
+    </ButtonContainer>
+  );
+};
+
+IconButton.Variant = IconButtonVariant;
+IconButton.Size = {
+  Small: 'small',
+  Medium: 'medium',
+} as const;
+
+export default IconButton;
+
+const getIconButtonColors = (variant: IconButtonVariant, toggled?: boolean): ButtonColors => {
+  switch (variant) {
+    case IconButton.Variant.Square:
+    case IconButtonVariant.Circle:
+    default:
+      return {
+        default: {
+          background: toggled ? colors.blueberry400 : undefined,
+          icon: toggled ? colors.frenchVanilla100 : colors.licorice200,
+        },
+        hover: {
+          background: toggled ? colors.blueberry500 : colors.soap300,
+          icon: toggled ? colors.frenchVanilla100 : colors.licorice500,
+        },
+        active: {
+          background: toggled ? colors.blueberry500 : colors.soap500,
+          icon: toggled ? colors.frenchVanilla100 : colors.licorice500,
+        },
+        focus: {
+          background: toggled ? colors.blueberry500 : colors.soap300,
+          icon: toggled ? colors.frenchVanilla100 : colors.licorice500,
+        },
+        disabled: {
+          background: toggled ? colors.blueberry100 : 'transparent',
+          icon: toggled ? colors.blueberry300 : colors.soap600,
+        },
+      };
+    case IconButtonVariant.SquareFilled:
+    case IconButtonVariant.CircleFilled:
+      return {
+        default: {
+          background: toggled ? colors.blueberry400 : colors.soap200,
+          icon: toggled ? colors.frenchVanilla100 : colors.licorice200,
+        },
+        hover: {
+          background: toggled ? colors.blueberry500 : colors.soap400,
+          icon: toggled ? colors.frenchVanilla100 : colors.licorice500,
+        },
+        active: {
+          background: toggled ? colors.blueberry500 : colors.soap500,
+          icon: toggled ? colors.frenchVanilla100 : colors.licorice500,
+        },
+        focus: {
+          background: toggled ? colors.blueberry500 : colors.soap400,
+          icon: toggled ? colors.frenchVanilla100 : colors.licorice500,
+        },
+        disabled: {
+          background: toggled ? colors.blueberry100 : colors.soap100,
+          icon: toggled ? colors.blueberry300 : colors.soap600,
+        },
+      };
+    case IconButtonVariant.Plain:
+      return {
+        default: {
+          icon: toggled ? colors.blueberry400 : colors.licorice200,
+        },
+        hover: {
+          icon: toggled ? colors.blueberry400 : colors.licorice500,
+        },
+        active: {
+          icon: toggled ? colors.blueberry400 : colors.licorice500,
+        },
+        focus: {
+          icon: toggled ? colors.blueberry400 : colors.licorice500,
+          focusRing: focusRing(2, 0),
+        },
+        disabled: {
+          icon: toggled ? colors.blueberry200 : colors.soap600,
+        },
+      };
+    case IconButtonVariant.Inverse:
+      return {
+        default: {
+          background: toggled ? colors.frenchVanilla100 : undefined,
+          icon: toggled ? colors.blueberry400 : colors.frenchVanilla100,
+        },
+        hover: {
+          background: toggled ? colors.frenchVanilla100 : 'rgba(0, 0, 0, 0.2)',
+          icon: toggled ? colors.blueberry400 : colors.frenchVanilla100,
+        },
+        active: {
+          background: toggled ? colors.frenchVanilla100 : 'rgba(0, 0, 0, 0.3)',
+          icon: toggled ? colors.blueberry400 : colors.frenchVanilla100,
+        },
+        focus: {
+          background: toggled ? colors.frenchVanilla100 : 'rgba(0, 0, 0, 0.2)',
+          icon: toggled ? colors.blueberry400 : colors.frenchVanilla100,
+          focusRing: focusRing(2, 2, true, false, 'currentColor', colors.frenchVanilla100),
+        },
+        disabled: {
+          background: toggled ? 'rgba(255,255,255,0.75)' : 'transparent',
+          icon: toggled ? colors.blueberry400 : 'rgba(255, 255, 255, 0.75)',
+        },
+      };
+    case IconButtonVariant.InverseFilled:
+      return {
+        default: {
+          background: toggled ? colors.frenchVanilla100 : 'rgba(0, 0, 0, 0.2)',
+          icon: toggled ? colors.blueberry400 : colors.frenchVanilla100,
+        },
+        hover: {
+          background: toggled ? colors.frenchVanilla100 : 'rgba(0, 0, 0, 0.3)',
+          icon: toggled ? colors.blueberry400 : colors.frenchVanilla100,
+        },
+        active: {
+          background: toggled ? colors.frenchVanilla100 : 'rgba(0, 0, 0, 0.4)',
+          icon: toggled ? colors.blueberry400 : colors.frenchVanilla100,
+        },
+        focus: {
+          background: toggled ? colors.frenchVanilla100 : 'rgba(0, 0, 0, 0.2)',
+          icon: toggled ? colors.blueberry400 : colors.frenchVanilla100,
+          focusRing: focusRing(2, 2, true, false, 'currentColor', colors.frenchVanilla100),
+        },
+        disabled: {
+          background: toggled ? 'rgba(255,255,255,0.75)' : 'rgba(0, 0, 0, 0.2)',
+          icon: toggled ? colors.blueberry400 : 'rgba(255, 255, 255, 0.75)',
+        },
+      };
   }
-);
-
-export default class IconButton extends React.Component<IconButtonProps> {
-  public static Variant = IconButtonVariant;
-  public static Size = IconButtonSize;
-
-  static defaultProps = {
-    variant: IconButtonVariant.Circle,
-    size: IconButtonSize.Medium,
-  } as const;
-
-  componentDidUpdate(prevProps: IconButtonProps) {
-    if (
-      prevProps.toggled !== this.props.toggled &&
-      typeof this.props.onToggleChange === 'function'
-    ) {
-      this.props.onToggleChange(this.props.toggled);
-    }
-  }
-
-  public render() {
-    // onToggleChange will generate a warning if spread over a <button>
-    const {
-      buttonRef,
-      size,
-      variant,
-      onToggleChange,
-      icon,
-      toggled,
-      children,
-      className,
-      ...elemProps
-    } = this.props;
-
-    return (
-      <ClassNames>
-        {({cx}) => (
-          <IconButtonCon
-            toggled={toggled}
-            ref={buttonRef}
-            variant={variant}
-            size={size}
-            aria-pressed={toggled}
-            className={cx(iconButtonIdentifier, className)}
-            {...elemProps}
-          >
-            {icon ? <SystemIcon icon={icon} /> : children}
-          </IconButtonCon>
-        )}
-      </ClassNames>
-    );
-  }
-}
+};
