+++ modules/button/react/lib/TextButton.tsx
@@ -1,74 +1,141 @@
 import * as React from 'react';
-import {ButtonBaseLabel, ButtonLabelIcon} from './ButtonBase';
-import {getButtonStyle} from './utils';
-import {styled} from '@workday/canvas-kit-labs-react-core';
-import isPropValid from '@emotion/is-prop-valid';
-import {ButtonSize, IconPosition, TextButtonVariant} from './types';
-import {BaseButtonProps} from './Button';
-import {textButtonStyles} from './ButtonStyles';
+import {type} from '@workday/canvas-kit-labs-react-core';
+import {focusRing} from '@workday/canvas-kit-react-common';
+import {colors, spacing, borderRadius} from '@workday/canvas-kit-react-core';
+import {CanvasSystemIcon} from '@workday/design-assets-types';
+import {TextButtonVariant, ButtonIconPosition, ButtonColors} from './types';
+import {ButtonContainer, ButtonLabelIcon, ButtonLabel} from './parts';
 
-export interface TextButtonProps extends BaseButtonProps<TextButtonVariant> {
+export interface TextButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+  /**
+   * The variant of the TextButton.
+   * @default TextButtonVariant.Default
+   */
+  variant?: TextButtonVariant;
+  /**
+   * The size of the TextButton.
+   * @default 'medium'
+   */
+  size?: 'small' | 'medium';
   /**
    * The position of the TextButton icon. Accepts `Left` or `Right`.
-   * @default IconPosition.Left
+   * @default ButtonIconPosition.Left
    */
-  iconPosition?: IconPosition;
+  iconPosition?: ButtonIconPosition;
+  /**
+   * The ref to the button that the styled component renders.
+   */
+  buttonRef?: React.Ref<HTMLButtonElement>;
+  /**
+   * The icon of the TextButton.
+   */
+  icon?: CanvasSystemIcon;
 }
 
-const TextButtonCon = styled('button', {
-  shouldForwardProp: prop => isPropValid(prop) && prop !== 'size',
-})<TextButtonProps>(
-  textButtonStyles.styles,
-  ({variant}) => getButtonStyle(textButtonStyles, variant),
-  ({size}) => {
-    const {sizes} = textButtonStyles.variants!;
-
-    switch (size) {
-      case ButtonSize.Large:
-      default:
-        return sizes.large;
-      case ButtonSize.Medium:
-      case ButtonSize.Small:
-        return sizes.small;
-    }
+const getTextButtonColors = (variant: TextButtonVariant): ButtonColors => {
+  switch (variant) {
+    case TextButtonVariant.Default:
+    case TextButtonVariant.AllCaps:
+    default:
+      return {
+        default: {
+          icon: colors.blueberry400,
+          label: colors.blueberry400,
+        },
+        hover: {
+          icon: colors.blueberry500,
+          label: colors.blueberry500,
+        },
+        active: {
+          icon: colors.blueberry500,
+          label: colors.blueberry500,
+        },
+        focus: {
+          icon: colors.blueberry500,
+          label: colors.blueberry500,
+          focusRing: focusRing(2, 0),
+        },
+        disabled: {
+          icon: 'rgba(8, 117, 225, 0.5)',
+          label: 'rgba(8, 117, 225, 0.5)',
+        },
+      };
+    case TextButtonVariant.Inverse:
+    case TextButtonVariant.InverseAllCaps:
+      return {
+        default: {
+          icon: colors.frenchVanilla100,
+          label: colors.frenchVanilla100,
+        },
+        hover: {},
+        active: {},
+        focus: {
+          focusRing: focusRing(2, 0, true, false, undefined, 'currentColor'),
+        },
+        disabled: {
+          icon: 'rgba(255, 255, 255, 0.5)',
+          label: 'rgba(255, 255, 255, 0.5)',
+        },
+      };
   }
-);
+};
 
-export default class TextButton extends React.Component<TextButtonProps> {
-  public static IconPosition = IconPosition;
-  public static Variant = TextButtonVariant;
-  public static Size = ButtonSize;
+const containerStyles = {
+  borderRadius: borderRadius.m,
+  border: '0',
+  padding: `0 ${spacing.xxs}`,
+  minWidth: 'auto',
+  '&:hover:not([disabled])': {textDecoration: 'underline'},
+};
 
-  static defaultProps = {
-    iconPosition: IconPosition.Left,
-    variant: TextButtonVariant.Default,
-    size: ButtonSize.Large,
-  };
+const TextButton = ({
+  variant = TextButtonVariant.Default,
+  size = 'medium',
+  iconPosition = ButtonIconPosition.Left,
+  buttonRef,
+  children,
+  icon,
+  ...elemProps
+}: TextButtonProps) => {
+  // Note: We don't use ButtonLabel because the label styles differ from other button types
+  const allContainerStyles =
+    variant === TextButtonVariant.AllCaps || variant === TextButtonVariant.InverseAllCaps
+      ? {
+          ...containerStyles,
+          ...type.variant.caps,
+          ...type.variant.button,
+          fontSize: size === 'medium' ? type.body.fontSize : undefined,
+          letterSpacing: '.5px',
+        }
+      : {
+          ...containerStyles,
+          fontSize: size === 'medium' ? type.body.fontSize : undefined,
+        };
 
-  public render() {
-    const {
-      buttonRef,
-      onClick,
-      children,
-      iconPosition,
-      size,
-      variant,
-      icon,
-      ...elemProps
-    } = this.props;
+  return (
+    <ButtonContainer
+      colors={getTextButtonColors(variant)}
+      ref={buttonRef}
+      size={size}
+      extraStyles={allContainerStyles}
+      {...elemProps}
+    >
+      {icon && iconPosition === ButtonIconPosition.Left && (
+        <ButtonLabelIcon size={size} iconPosition={iconPosition} icon={icon} />
+      )}
+      <ButtonLabel>{children}</ButtonLabel>
+      {icon && iconPosition === ButtonIconPosition.Right && (
+        <ButtonLabelIcon size={size} iconPosition={iconPosition} icon={icon} />
+      )}
+    </ButtonContainer>
+  );
+};
 
-    return (
-      <TextButtonCon onClick={onClick} ref={buttonRef} size={size} variant={variant} {...elemProps}>
-        {icon && iconPosition === IconPosition.Left && (
-          <ButtonLabelIcon size={size} iconPosition={iconPosition} icon={icon} />
-        )}
-        <ButtonBaseLabel size={size} variant={variant}>
-          {children}
-        </ButtonBaseLabel>
-        {icon && iconPosition === IconPosition.Right && (
-          <ButtonLabelIcon size={size} iconPosition={iconPosition} icon={icon} />
-        )}
-      </TextButtonCon>
-    );
-  }
-}
+TextButton.IconPosition = ButtonIconPosition;
+TextButton.Variant = TextButtonVariant;
+TextButton.Size = {
+  Small: 'small',
+  Medium: 'medium',
+} as const;
+
+export default TextButton;
