+++ modules/button/react/lib/parts/ButtonLabelIcon.tsx
