+++ modules/button/react/lib/types.ts
@@ -1,41 +1,31 @@
-/**
- * The different button sizes.
- */
-export enum ButtonSize {
-  Small = 'small',
-  Medium = 'medium',
-  Large = 'large',
-}
-
-export enum IconButtonSize {
-  Small = 'small',
-  Medium = 'medium',
-}
-
-// TODO (beta button): consolidate all these button types
-/**
- * The different button types.
- */
+import {CSSObject} from '@emotion/core';
 
 /**
- * @deprecated These type are deprecated along with deprecated_Button component
+ * Standard button Button
  */
-export enum DeprecatedButtonVariant {
-  Primary = 'deprecatedPrimary',
-  Secondary = 'deprecatedSecondary',
-  Delete = 'deprecatedDelete',
-}
-
 export enum ButtonVariant {
   Primary = 'primary',
-  Secondary = 'betaSecondary',
-  Delete = 'delete',
-  Highlight = 'highlight',
+  Secondary = 'secondary',
   OutlinePrimary = 'outlinePrimary',
   OutlineSecondary = 'outlineSecondary',
   OutlineInverse = 'outlineInverse',
 }
+export enum ButtonIconPosition {
+  Left = 'iconPositionLeft',
+  Right = 'iconPositionRight',
+}
 
+/**
+ * Dropdown Button
+ */
+export enum DropdownButtonVariant {
+  Primary = 'primary',
+  Secondary = 'secondary',
+}
+
+/**
+ * Icon Button
+ */
 export enum IconButtonVariant {
   Square = 'square',
   SquareFilled = 'squareFilled',
@@ -47,13 +37,8 @@ export enum IconButtonVariant {
 }
 
 /**
- * The different icon positions.
+ * Text Button
  */
-export enum IconPosition {
-  Left = 'iconPositionLeft',
-  Right = 'iconPositionRight',
-}
-
 export enum TextButtonVariant {
   Default = 'text',
   Inverse = 'textInverse',
@@ -61,8 +46,40 @@ export enum TextButtonVariant {
   InverseAllCaps = 'textInverseAllCaps',
 }
 
+/**
+ * Old orange buttons
+ * @deprecated These type are deprecated along with deprecated_Button component
+ */
+export enum DeprecatedButtonVariant {
+  Primary = 'deprecatedPrimary',
+  Secondary = 'deprecatedSecondary',
+  Delete = 'deprecatedDelete',
+}
+
 export type AllButtonVariants =
-  | DeprecatedButtonVariant
   | ButtonVariant
+  | DropdownButtonVariant
   | TextButtonVariant
-  | IconButtonVariant;
+  | IconButtonVariant
+  | DeprecatedButtonVariant;
+
+/**
+ * The object used for passing in colors to the ButtonContainer
+ */
+export interface ButtonStateColors {
+  background?: string;
+  border?: string;
+  icon?: string;
+  iconFill?: boolean;
+  label?: string;
+  labelData?: string;
+}
+export interface ButtonColors {
+  default: ButtonStateColors;
+  hover: ButtonStateColors;
+  active: ButtonStateColors;
+  focus: ButtonStateColors & {
+    focusRing?: CSSObject;
+  };
+  disabled: ButtonStateColors;
+}
