+++ modules/button/react/package.json
@@ -1,7 +1,8 @@
+
 {
-  "name": "@workday/canvas-kit-react-button",
-  "version": "3.4.0",
-  "description": "A canvas-styled react button",
+  "name": "@workday/canvas-kit-labs-react-button",
+  "version": "0.0.0",
+  "description": "A collection of various Canvas Buttons",
   "author": "Workday, Inc. (https://www.workday.com)",
   "license": "Apache-2.0",
   "main": "dist/commonjs/index.js",
@@ -10,7 +11,7 @@
   "types": "dist/es6/index.d.ts",
   "repository": {
     "type": "git",
-    "url": "http://github.com/Workday/canvas-kit/tree/master/modules/button/react"
+    "url": "https://github.com/Workday/canvas-kit/tree/master/modules/_labs/button/react"
   },
   "files": [
     "dist/",
@@ -19,13 +20,11 @@
   ],
   "scripts": {
     "watch": "yarn build:es6 -w",
-    "test": "echo \"Error: no test specified\" && exit 1",
     "clean": "rimraf dist && rimraf .build-info && mkdirp dist",
-    "build:cjs": "babel --no-babelrc --plugins babel-plugin-transform-es2015-modules-commonjs dist/es6 --out-dir dist/commonjs  --copy-files",
+    "build:cjs": "tsc -p tsconfig.cjs.json",
     "build:es6": "tsc -p tsconfig.es6.json",
-    "build:babel": "babel --no-babelrc --plugins babel-plugin-emotion dist --out-dir dist",
     "build:rebuild": "npm-run-all clean build",
-    "build": "npm-run-all --sequential build:es6 build:babel build:cjs",
+    "build": "npm-run-all --parallel build:cjs build:es6",
     "depcheck": "node ../../../utils/check-dependencies-exist.js"
   },
   "keywords": [
