+++ modules/button/react/tsconfig.es6.json
@@ -1,3 +1,4 @@
+
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
